WebUI change

+++ panel/audio/convert.sh
@@ -1,15 +1,18 @@
 #!/bin/sh
 SOX="sox"
 SRC="/var/spool/asterisk/recording"
-FLIST="abs-tcmessage abs-tcrmessage"
+FLIST=$*
 TARGET="audio"
 
-for i in $FLIST
-do
-  rm -f $i.mp3
-done
+if [ "$FLIST" != "" ];
+then
+  for i in $FLIST
+  do
+    rm -f $i.mp3
+  done
 
-for i in $FLIST
-do
-  $SOX $SRC/$i.wav $TARGET/$i.mp3
-done
+  for i in $FLIST
+  do
+    $SOX $SRC/$i.gsm $TARGET/$i.mp3
+  done
+fi
